+++ tests/unit/fsp/conftest.py
@@ -82,9 +82,17 @@ def mock_fsp_pst_subarray_device() -> unittest.mock.Mock:
 
 
 @pytest.fixture()
+def mock_hps_fsp_corr_controller() -> unittest.mock.Mock:
+    builder = MockDeviceBuilder()
+    builder.add_command("SetFunctionMode", None)
+    return builder()
+
+
+@pytest.fixture()
 def initial_mocks(
     mock_fsp_corr_subarray_device: unittest.mock.Mock,
     mock_fsp_pst_subarray_device: unittest.mock.Mock,
+    mock_hps_fsp_corr_controller: unittest.mock.Mock,
 ) -> dict[str, unittest.mock.Mock]:
     """
     Return a dictionary of device proxy mocks to pre-register.
@@ -101,4 +109,5 @@ def initial_mocks(
         mocks[
             f"mid_csp_cbf/fspPstSubarray/01_{sub_id:02}"
         ] = mock_fsp_pst_subarray_device
+    mocks["talondx-001/fsp-app/fsp-controller"] = mock_hps_fsp_corr_controller
     return mocks