+++ tests/unit/fsp_corr_subarray/FspCorrSubarrayDevice_test.py
@@ -109,7 +109,7 @@ class TestFspCorrSubarray:
                              events from the device under test.
         """
         # Set the DUT to AdminMode.ONLINE and DevState.ON
-        device_under_test.simulationMode == SimulationMode.FALSE
+        device_under_test.simulationMode = SimulationMode.FALSE
         device_under_test.adminMode = AdminMode.ONLINE
         assert_that(event_tracer).within_timeout(
             test_utils.EVENT_TIMEOUT