+++ CHANGELOG.rst
@@ -25,6 +25,7 @@ UNRELEASED CHANGES
   * Controller OpState is ON if one or more LRUs are powered on, and only OFF once all LRUs are powered off and Off command is fully successful
   * wait_for_blocking_results allows partial success of blocking LRCs
   * Configurable HPS device timeouts in charts: CbfController, FspCorrSubarray, SlimLink, Vcc
+* CIP-2899: Refactored adminMode push event to be changed after communication has been established
 
 1.1.0-rc.1
 **********